elements: merge view + component into dual-mode <patchwork-view>; rename legacy impl
Collapse the standalone <patchwork-component> element into <patchwork-view>:
the wrapper picks component mode (`component`/`url` attrs) or legacy mode
(`doc-url`/`tool-id` attrs, delegated to <patchwork-view-legacy>). Inline
the JSX intrinsic augmentations into index.ts as a discriminated union so
mixing the two modes is a compile-time error, and drop the hand-written
elements.d.ts. Bootloader no longer registers <patchwork-component>; the
single registerPatchworkViewElement() call registers both names. Update
changesets accordingly.
