+++ .changeset/patchwork-bootloader-providers.md
@@ -7,9 +7,11 @@ Wire the new `@inkandswitch/patchwork-providers` element stack into
 
 - Register and mount `<repo-provider>` (backed by the booted repo) and a
   top-level `<fallback-provider>` around the configured root element so
-  descendant `<patchwork-view>` / `<patchwork-component>` elements can
+  descendant `<patchwork-view>` / `<patchwork-view-legacy>` elements can
   resolve their repo via the request/respond protocol.
-- Register the new `<patchwork-component>` element in addition to
-  `<patchwork-view>`, and drop the `{ repo }` argument from
-  `registerPatchworkViewElement` (the repo now comes from the provider).
+- The single `registerPatchworkViewElement()` call now also registers
+  `<patchwork-view-legacy>` (the wrapper's delegation target), so no
+  separate registration is needed.
+- Drop the `{ repo }` argument from `registerPatchworkViewLegacyElement`
+  (the repo now comes from the provider).
 - Add `@inkandswitch/patchwork-providers` as a workspace dependency.
