+++ core/elements/src/patchwork-view.ts
@@ -1,30 +1,30 @@
+import type { AutomergeUrl } from "@automerge/automerge-repo";
 import {
-  type AutomergeUrl,
-  type DocHandle,
-  type DocHandleChangePayload,
-  type Repo,
-} from "@automerge/automerge-repo";
-import {
-  getType,
-  type HasPatchworkMetadata,
-} from "@inkandswitch/patchwork-filesystem";
-import {
-  getFallbackTool,
   getRegistry,
   isLoadablePlugin,
-  type LoadedTool,
+  type LoadablePlugin,
+  type LoadedPlugin,
+  type PluginDescription,
 } from "@inkandswitch/patchwork-plugins";
-import { request } from "@inkandswitch/patchwork-providers";
-import debug from "debug";
-import { MountedEvent, NoToolEvent, UnmountedEvent } from "./events.js";
+import { MountedEvent, UnmountedEvent } from "./events.js";
+import { registerPatchworkViewLegacyElement } from "./patchwork-view-legacy.js";
 
-const log = debug("patchwork:elements:view");
+export type ComponentRender = (element: HTMLElement) => () => void;
 
-import type { initializeAutomergeRepoKeyhive } from "@automerge/automerge-repo-keyhive";
+export type ComponentDescription = PluginDescription & {
+  id: string;
+  type: "patchwork:component";
+  name: string;
+  icon?: string;
+  tags?: string[];
+};
 
-type AutomergeRepoKeyhive = Awaited<
-  ReturnType<typeof initializeAutomergeRepoKeyhive>
+export type LoadedComponent = LoadedPlugin<
+  ComponentDescription,
+  ComponentRender
 >;
+
+export type Component = LoadablePlugin<ComponentDescription, ComponentRender>;
 
 const State = {
   none: "none",
@@ -32,22 +32,42 @@ const State = {
   rendering: "rendering",
   unable: "unable",
   rendered: "rendered",
-  fallback: "fallback",
   error: "error",
 } as const;
 
 type State = (typeof State)[keyof typeof State];
 
+const Mode = {
+  idle: "idle",
+  component: "component",
+  legacy: "legacy",
+} as const;
+
+type Mode = (typeof Mode)[keyof typeof Mode];
+
+const ATTRS = {
+  component: "component",
+  url: "url",
+  docUrl: "doc-url",
+  toolId: "tool-id",
+} as const;
+
 export type RegisterPatchworkViewElementParams = {
   name?: string;
-  hive?: AutomergeRepoKeyhive;
 };
 
+/**
+ * `<patchwork-view>` operates in one of two modes based on its
+ * attributes: **component mode** (`component`) mounts a registered
+ * `patchwork:component` plugin in place; **legacy mode** (`doc-url` /
+ * `tool-id`, no `component`) delegates to an inner
+ * `<patchwork-view-legacy>`. `component` wins if both are set.
+ */
 export type PatchworkViewElement = HTMLElement & {
-  repo: Repo;
-  hive?: AutomergeRepoKeyhive;
-  docUrl?: AutomergeUrl;
-  toolId?: string;
+  component?: string | null;
+  url?: AutomergeUrl | null;
+  docUrl?: AutomergeUrl | null;
+  toolId?: string | null;
 };
 
 export function registerPatchworkViewElement(
@@ -55,73 +75,86 @@ export function registerPatchworkViewElement(
 ) {
   const name = params.name ?? "patchwork-view";
 
-  if (customElements.get(name)) return;
+  // Also registers `<patchwork-view-legacy>` (the wrapper's delegation
+  // target). Both register functions are idempotent.
+  registerPatchworkViewLegacyElement();
 
-  const attrs = {
-    docUrl: "doc-url",
-    toolId: "tool-id",
-  };
+  if (customElements.get(name)) return;
 
   customElements.define(
     name,
     class PatchworkViewElement extends HTMLElement {
-      repo!: Repo;
-      hive = params.hive;
-      // attributes, if these change it's new game +
-      #docUrl: AutomergeUrl | null = null;
-      #toolId: string | null = null;
-      #handle: DocHandle<HasPatchworkMetadata> | null = null;
-      #tool: LoadedTool | null = null;
-      #state: State = State.none;
-      #requestedToolImports = new Set<string>();
-      #initEpoch = 0;
+      #mode: Mode = Mode.idle;
       #capturedParent: Element | null = null;
 
-      get docUrl() {
-        return this.#docUrl;
+      #component: string | null = null;
+      #url: AutomergeUrl | null = null;
+      #loaded: LoadedComponent | null = null;
+      #state: State = State.none;
+      #initEpoch = 0;
+      #teardowns = new Set<() => unknown | Promise<void>>();
+
+      #legacyChild: HTMLElement | null = null;
+
+      get component() {
+        return this.#component;
+      }
+
+      set component(id: string | null) {
+        if (this.#component === id) return;
+        this.#component = id;
+        const attr = this.getAttribute(ATTRS.component);
+        if (attr == id) return;
+        if (id) this.setAttribute(ATTRS.component, id);
+        else this.removeAttribute(ATTRS.component);
+      }
+
+      get url() {
+        return this.#url;
+      }
+
+      set url(url: AutomergeUrl | null) {
+        if (this.#url === url) return;
+        this.#url = url;
+        const attr = this.getAttribute(ATTRS.url);
+        if (attr == url) return;
+        if (url) this.setAttribute(ATTRS.url, url);
+        else this.removeAttribute(ATTRS.url);
+      }
+
+      get docUrl(): AutomergeUrl | null {
+        return this.getAttribute(ATTRS.docUrl) as AutomergeUrl | null;
       }
 
       set docUrl(url: AutomergeUrl | null) {
-        if (this.#docUrl === url) return;
-        this.#docUrl = url;
-        const attr = this.getAttribute(attrs.docUrl);
-        if (attr == url) return;
-        if (url) {
-          this.setAttribute(attrs.docUrl, url);
-        } else {
-          this.removeAttribute(attrs.docUrl);
-        }
+        if (this.docUrl === url) return;
+        if (url) this.setAttribute(ATTRS.docUrl, url);
+        else this.removeAttribute(ATTRS.docUrl);
       }
 
-      get toolId() {
-        return this.#toolId;
+      get toolId(): string | null {
+        return this.getAttribute(ATTRS.toolId);
       }
 
       set toolId(id: string | null) {
-        if (this.#toolId === id) return;
-        this.#toolId = id;
-        const attr = this.getAttribute(attrs.toolId);
-        if (attr == id) return;
-        if (id) {
-          this.setAttribute(attrs.toolId, id);
-        } else {
-          this.removeAttribute(attrs.toolId);
-        }
+        if (this.toolId === id) return;
+        if (id) this.setAttribute(ATTRS.toolId, id);
+        else this.removeAttribute(ATTRS.toolId);
       }
 
       static get observedAttributes() {
-        return [attrs.docUrl, attrs.toolId];
+        return [ATTRS.component, ATTRS.url, ATTRS.docUrl, ATTRS.toolId];
       }
 
       connectedCallback() {
         this.#capturedParent = this.parentElement;
-        this.docUrl = this.getAttribute(attrs.docUrl) as AutomergeUrl;
-        this.toolId = this.getAttribute(attrs.toolId);
-        this.#init();
+        this.#component = this.getAttribute(ATTRS.component);
+        this.#url = this.getAttribute(ATTRS.url) as AutomergeUrl | null;
+        this.#syncMode();
       }
 
       disconnectedCallback() {
-        this.#teardown();
+        void this.#teardown();
       }
 
       connectedMoveCallback() {
@@ -129,166 +162,162 @@ export function registerPatchworkViewElement(
       }
 
       attributeChangedCallback(
-        name: string,
+        attrName: string,
         old: string | null,
         val: string | null
       ) {
         if (old === val) return;
 
-        if (name === attrs.toolId) {
-          this.#toolId = val;
-          this.#teardown().then(() => this.#init());
+        if (attrName === ATTRS.component) this.#component = val;
+        else if (attrName === ATTRS.url) this.#url = val as AutomergeUrl | null;
+
+        const targetMode = this.#pickMode();
+        if (targetMode !== this.#mode) {
+          void this.#teardown().then(() => this.#syncMode());
+          return;
         }
 
-        if (name === attrs.docUrl) {
-          this.#docUrl = val as AutomergeUrl;
-          this.#teardown().then(() => this.#init());
+        if (this.#mode === Mode.legacy && this.#legacyChild) {
+          if (attrName === ATTRS.docUrl || attrName === ATTRS.toolId) {
+            if (val !== null) this.#legacyChild.setAttribute(attrName, val);
+            else this.#legacyChild.removeAttribute(attrName);
+          }
+          return;
+        }
+
+        if (this.#mode === Mode.component) {
+          if (attrName === ATTRS.component || attrName === ATTRS.url) {
+            void this.#teardown().then(() => this.#syncMode());
+          }
         }
       }
 
-      #onDocChange = (
-        payload: DocHandleChangePayload<HasPatchworkMetadata>
-      ) => {
-        const { before, after } = payload.patchInfo;
-
-        if (getType(before) != getType(after)) {
-          this.#teardown().then(() => this.#init());
+      #pickMode(): Mode {
+        if (this.hasAttribute(ATTRS.component)) return Mode.component;
+        if (this.hasAttribute(ATTRS.docUrl) || this.hasAttribute(ATTRS.toolId)) {
+          return Mode.legacy;
         }
-      };
+        return Mode.idle;
+      }
 
-      #init = async () => {
-        const toolRegistry = getRegistry("patchwork:tool");
-        if (this.#state != State.none) {
+      #syncMode() {
+        const target = this.#pickMode();
+        this.#mode = target;
+
+        if (target === Mode.idle) return;
+
+        if (target === Mode.legacy) {
+          this.#initLegacy();
           return;
         }
 
-        if (!this.docUrl) {
-          return;
-        }
+        // Component-only: legacy mode leaves `display` alone so sites'
+        // CSS rules on `patchwork-view` (e.g. `display: block`) apply.
+        if (!this.style.display) this.style.display = "contents";
+        this.#initComponent();
+      }
+
+      #initLegacy() {
+        if (this.#legacyChild) return;
+        const child = document.createElement("patchwork-view-legacy");
+        const docUrl = this.getAttribute(ATTRS.docUrl);
+        const toolId = this.getAttribute(ATTRS.toolId);
+        if (docUrl) child.setAttribute(ATTRS.docUrl, docUrl);
+        if (toolId) child.setAttribute(ATTRS.toolId, toolId);
+        this.appendChild(child);
+        this.#legacyChild = child;
+      }
+
+      #initComponent = async () => {
+        if (this.#state != State.none) return;
+        if (!this.component) return;
+
+        const registry = getRegistry<ComponentDescription>(
+          "patchwork:component"
+        );
 
         const epoch = ++this.#initEpoch;
         this.#state = State.initializing;
 
-        const removeAddedListener = toolRegistry.on(
+        const removeAddedListener = registry.on(
           "registered",
-          async (addedTool) => {
-            const toolId = addedTool.id;
-            const isChosenTool = toolId == this.toolId;
-            if (this.#handle) {
-              this.#fallbackId = getFallbackTool(this.#handle.doc())?.id;
-            }
-            const isFallbackTool = toolId == this.#fallbackId;
-
-            if (isChosenTool || isFallbackTool) {
-              if (isLoadablePlugin(addedTool)) {
-                // if it's not loaded, load it now
-                toolRegistry.load(addedTool.id);
-              }
+          async (added) => {
+            if (added.id !== this.component) return;
+            if (isLoadablePlugin(added)) {
+              registry.load(added.id);
             }
           }
         );
 
-        const removeLoadedListener = toolRegistry.on(
-          "loaded",
-          async (loadedTool) => {
-            const toolId = loadedTool.id;
-            const isChosenTool = toolId == this.toolId;
-            const isFallbackTool = toolId == this.#fallbackId;
+        const removeLoadedListener = registry.on("loaded", async (loaded) => {
+          if (loaded.id !== this.component) return;
 
-            if (isChosenTool || isFallbackTool) {
-              if (this.#state == "unable" || this.#state == "initializing") {
-                this.#queueRender();
-              }
+          if (this.#state == "unable" || this.#state == "initializing") {
+            this.#queueRender();
+            return;
+          }
 
-              if (
-                ((this.#state == "error" || this.#state == "rendered") &&
-                  isChosenTool) ||
-                (this.#state == "fallback" && isFallbackTool)
-              ) {
-                if (loadedTool.importUrl !== this.#tool?.importUrl) {
-                  await this.#teardown();
-                  this.#init();
-                }
-              }
+          // Hot reload: a newer importUrl re-mounts.
+          if (this.#state == "error" || this.#state == "rendered") {
+            if (loaded.importUrl !== this.#loaded?.importUrl) {
+              await this.#teardown();
+              this.#syncMode();
             }
           }
-        );
+        });
 
         this.#teardowns.add(() => {
           removeAddedListener();
           removeLoadedListener();
         });
 
-        const repo = await request<Repo>(this, "patchwork:repo");
         if (epoch !== this.#initEpoch) return;
-        if (!repo) {
-          this.#state = State.unable;
-          this.#displayError(
-            `no \`patchwork:repo\` provider in the DOM ancestry of <${name}>.`
-          );
-          return;
-        }
-        this.repo = repo;
-
-        let handle: DocHandle<HasPatchworkMetadata>;
-        try {
-          handle = await repo.find<HasPatchworkMetadata>(this.docUrl!);
-        } catch (err) {
-          // If teardown ran while we were awaiting, the listener cleanup
-          // already happened — silently abort.
-          if (epoch !== this.#initEpoch) return;
-          throw err;
-        }
-
-        // Teardown ran during the await — abort before clobbering state.
-        if (epoch !== this.#initEpoch) return;
-
-        this.#handle = handle;
-        this.#handle.on("change", this.#onDocChange);
-        this.#teardowns.add(() =>
-          this.#handle!.off("change", this.#onDocChange)
-        );
 
         this.#queueRender();
       };
 
-      #teardowns = new Set<() => unknown | Promise<void>>();
-
       async #teardown() {
-        if (this.#state == State.none) return;
-
-        // Capture the mount-event payload (if any) before clearing state,
-        // so the unmount echoes the same {url, toolId} we dispatched at
-        // mount. `rendered`/`fallback` are exactly the states in which
-        // `#render` dispatched a MountedEvent (it sets `#tool` only after
-        // the tool's module call succeeds).
-        const wasMounted =
-          this.#state == State.rendered || this.#state == State.fallback;
-        const mountedUrl = this.#handle?.url;
-        const mountedToolId = this.#tool?.id;
-
-        this.#initEpoch++;
-
-        for (const fn of this.#teardowns) {
-          await fn?.();
+        if (this.#mode === Mode.legacy) {
+          if (this.#legacyChild) {
+            this.#legacyChild.remove();
+            this.#legacyChild = null;
+          }
+          this.#mode = Mode.idle;
+          return;
         }
 
-        this.#teardowns.clear();
-        this.#handle = null;
-        this.#tool = null;
-        this.#requestedToolImports.clear();
-        this.textContent = "";
-        this.#state = State.none;
+        if (this.#mode === Mode.component) {
+          if (this.#state == State.none) {
+            this.#mode = Mode.idle;
+            return;
+          }
 
-        if (wasMounted && mountedUrl && mountedToolId) {
-          this.#dispatchUnmount(
-            new UnmountedEvent({ url: mountedUrl, toolId: mountedToolId })
-          );
+          // Only `rendered` reached the `MountedEvent` dispatch.
+          const wasMounted = this.#state == State.rendered;
+          const mountedComponentId = this.#loaded?.id;
+
+          this.#initEpoch++;
+
+          for (const fn of this.#teardowns) {
+            await fn?.();
+          }
+
+          this.#teardowns.clear();
+          this.#loaded = null;
+          this.#state = State.none;
+
+          if (wasMounted && mountedComponentId) {
+            this.#dispatchUnmount(
+              new UnmountedEvent({ componentId: mountedComponentId })
+            );
+          }
         }
+
+        this.#mode = Mode.idle;
       }
 
-      // Detached elements have no parent path, so `bubbles: true` is a
-      // no-op; fall back to the closest still-connected ancestor.
+      // Bubbling is a no-op when detached; fall back to the closest
+      // still-connected ancestor.
       #dispatchUnmount(event: UnmountedEvent) {
         if (this.isConnected) {
           this.dispatchEvent(event);
@@ -303,171 +332,61 @@ export function registerPatchworkViewElement(
         if (this.#state == "none") return;
         if (this.#state == "rendering") return;
         this.#state = "rendering";
-        queueMicrotask(() => this.#render());
+        queueMicrotask(() => this.#renderComponent());
       }
 
-      #fallbackId: string | undefined;
-
-      #render() {
+      #renderComponent() {
         if (this.#state != "rendering") return;
-        if (!this.docUrl || !this.#handle) {
+        if (!this.component) {
           this.#state = "unable";
           return;
         }
 
-        // Clear any previous content and error styles
-        this.#resetDisplay();
+        const componentId = this.component;
+        const registry = getRegistry<LoadedComponent>("patchwork:component");
+        this.#loaded = registry.get(componentId) ?? null;
 
-        const doc = this.#handle.doc();
-        this.#fallbackId = getFallbackTool(doc)?.id;
-        const fallingBack = !this.toolId;
-        const toolId = this.toolId || this.#fallbackId;
-
-        if (fallingBack) {
-          console.warn(`falling back to default tool for ${this.#docUrl}`);
+        if (!this.#loaded) {
+          this.#state = "unable";
+          console.warn(
+            `patchwork-view: no component registered with id "${componentId}"`
+          );
+          return;
         }
 
-        if (!toolId) {
-          this.#state = "unable";
-
-          // Check if the document is missing @patchwork metadata
-          const hasPatchworkMetadata = doc && "@patchwork" in doc;
-          if (!hasPatchworkMetadata) {
+        if (!this.#loaded.module) {
+          registry.load(this.#loaded.id);
+          if (registry.isLoading(this.#loaded.id)) {
+            this.#state = "unable";
+            console.info(
+              `patchwork-view: loading component "${componentId}"`
+            );
+          } else {
+            this.#state = "unable";
             console.warn(
-              `Document ${this.#docUrl} is missing @patchwork metadata`
+              `patchwork-view: failed to load component "${componentId}"`
             );
-            this.#displayError(
-              `This document is missing @patchwork metadata and cannot be opened.`
-            );
-          } else {
-            console.warn(`no tool for ${this.#docUrl}`);
-            this.#displayError(
-              `I couldn't find a tool to open ${this.#docUrl}.`
-            );
-          }
-          return;
-        }
-
-        this.#tool =
-          getRegistry<LoadedTool>("patchwork:tool").get(toolId) ?? null;
-
-        if (!this.#tool) {
-          this.#notool();
-        }
-
-        if (!this.#tool) {
-          this.#state = "unable";
-          this.#displayError(`I couldn't find the tool with id ${toolId}.`);
-          return;
-        }
-
-        if (!this.#tool.module) {
-          const toolRegistry = getRegistry("patchwork:tool");
-          toolRegistry.load(this.#tool.id);
-          if (toolRegistry.isLoading(this.#tool.id)) {
-            this.#state = "unable";
-            log(`loading ${toolId}`);
-            this.#displayLoading(toolId);
-          } else {
-            this.#state = "unable";
-            this.#displayError(`I couldn't load the tool with id ${toolId}.`);
           }
           return;
         }
 
         try {
-          const cleanup = this.#tool.module(this.#handle, this);
+          const cleanup = this.#loaded.module(this);
           if (typeof cleanup === "function") {
             this.#teardowns.add(cleanup);
           } else {
-            console.warn(`return a cleanup function from ${toolId}`);
+            console.warn(`return a cleanup function from ${componentId}`);
           }
-          this.#state = fallingBack ? "fallback" : "rendered";
-          this.dispatchEvent(new MountedEvent({ url: this.docUrl, toolId }));
+          this.#state = "rendered";
+          this.dispatchEvent(new MountedEvent({ componentId }));
         } catch (error) {
-          this.append(
-            Object.assign(document.createElement("div"), {
-              innerHTML: /* html */ `
-                <p>oh no!</p>
-                <details>
-                  <summary>${(error as Error).message ?? error}</summary>
-                  <pre style="white-space: pre-wrap;">${(error as Error).stack ?? ""}</pre>
-                </details>
-              `,
-            })
+          console.error(
+            `patchwork-view: component "${componentId}" threw during mount`,
+            error
           );
-          console.error(error);
-
           this.#state = "error";
         }
       }
-
-      #displayLoading = (toolId: string) => {
-        const div = document.createElement("div");
-        div.style.display = "flex";
-        div.style.alignItems = "center";
-        div.style.justifyContent = "center";
-        div.style.height = "100%";
-        div.innerHTML = /* html */ `
-          <style>
-            @keyframes pw-loading-spin {
-              to { transform: rotate(360deg); }
-            }
-          </style>
-          <div style="display: flex; flex-direction: column; align-items: center; gap: 8px;">
-            <div style="
-              width: 24px;
-              height: 24px;
-              border: 3px solid #e0e0e0;
-              border-top-color: #888;
-              border-radius: 50%;
-              animation: pw-loading-spin 0.8s linear infinite;
-            "></div>
-            <div style="font-size: 12px; color: #888;">loading ${toolId}</div>
-          </div>
-        `;
-        this.append(div);
-      };
-
-      #displayError = (error: string) => {
-        const div = document.createElement("div");
-        div.style.display = "flex";
-        div.style.alignItems = "center";
-        div.style.justifyContent = "center";
-        // wait a second then face in over half a second
-        div.style.transition = "opacity 2s linear 2s";
-        div.style.opacity = "0";
-        div.innerHTML = /* html */ `
-          <details style="display: flex"><summary></summary><pre style="white-space: pre-wrap;"><code>${error}</code></pre></details>
-        `;
-        this.append(div);
-        setTimeout(() => {
-          div.style.opacity = "1";
-        });
-      };
-
-      #notool() {
-        if (!this.docUrl || !this.#handle) return;
-        const suggestedImportUrl =
-          this.#handle.doc()?.["@patchwork"]?.suggestedImportUrl;
-        if (
-          !suggestedImportUrl ||
-          this.#requestedToolImports.has(suggestedImportUrl)
-        )
-          return;
-        this.#requestedToolImports.add(suggestedImportUrl);
-        let url = this.docUrl;
-
-        log("dispatching patchwork:no-tool for", this.docUrl);
-        this.dispatchEvent(new NoToolEvent({ url }));
-      }
-
-      #resetDisplay = () => {
-        this.replaceChildren();
-        this.style.display = "";
-        this.style.alignItems = "";
-        this.style.justifyContent = "";
-      };
     }
   );
 }
